pyproject.toml
[project]
name = "telegrambot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elasticsearch>=9.0.2",
    "fastapi>=0.115.14",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.1",
    "uvicorn>=0.34.3",
]
